Ints to doubles in entity, player has minor changes

+++ src/entity/Entity.java
@@ -7,14 +7,14 @@ import java.lang.Runnable;
 
 public abstract class Entity implements Runnable {
 	
-	protected int xPos;
-	protected int yPos;
+	protected double xPos;
+	protected double yPos;
 
 	protected int width;
 	protected int height;
 
-	protected int xSpeed;
-	protected int ySpeed;
+	protected double xSpeed;
+	protected double ySpeed;
 
     protected List<Entity> collisions;
 
@@ -60,19 +60,19 @@ public abstract class Entity implements Runnable {
         collisions = new ArrayList<Entity>();
     }
 
-	public int get_x() {
+	public double get_x() {
 		return xPos;
 	}
 
-	public void set_x(int xPos) {
+	public void set_x(double xPos) {
 		this.xPos = xPos;
 	}
 
-	public int get_y() {
+	public double get_y() {
 		return yPos;
 	}
 
-	public void set_y(int yPos) {
+	public void set_y(double yPos) {
 		this.yPos = yPos;
 	}
